Fix minor issues when building for 64-bit Windows

+++ src/DirManager.cpp
@@ -334,7 +334,7 @@ void DirManager::RecursivelyRemove(const FilePaths& filePathArray, int count, in
          }
       }
       if (progress)
-         progress->Update(ii + bias, count);
+         progress->Update(ii + bias, (size_t) count);
    }
 }
 
@@ -690,7 +690,7 @@ DirManager::ProjectSetter::Impl::Impl(
       ProgressDialog progress(XO("Progress"),
          XO("Saving project data files"));
 
-      int total = dirManager.mBlockFileHash.size();
+      size_t total = dirManager.mBlockFileHash.size();
 
       bool link = moving;
       for (const auto &pair : dirManager.mBlockFileHash) {
